+++ app/forms/__init__.py
@@ -1,6 +1,3 @@
-from .base import *
-from .login import *
-from .register import * 
-from .profile import * 
-from .stock import *
-from .shares import *
+from .campaign import *
+
+from .forms import *
