+++ app/routes/__init__.py
@@ -1,6 +1,5 @@
-from .debug import *
-from .auth import *
-from .home import *
-from .profile import *
-from .stock import *
-from .requests import *
+from .login import *
+from .mail import *
+from .campaign import *
+from .portfolio import *
+from .landing import *
