--- app/routes/utils/__init__.py
@@ -1,4 +1,0 @@
-from .login_required import *
-from .is_trader import *
-from .is_broker import *
-from .normalize import *
