+++ app/templates/campaign.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>New Campaign Form</title>
+</head>
+<body>
+
+  <h1>New Campaign Form</h1>
+
+  <form method="POST" action="{{ url_for('index') }}">
+      {{ form.hidden_tag() }}
+
+      <div>
+          {{ form.campaignName_Form.label }}
+          {{ form.campaignName_Form(class="newCampaignForm") }}
+          {% for error in form.campaignName_Form.errors %}
+              <span style="color: red;">{{ error }}</span>
+          {% endfor %}
+      </div>
+
+      <div>
+          {{ form.links_Form.label }}
+          {{ form.links_Form(class="newCampaignForm") }}
+          {% for error in form.links_Form.errors %}
+              <span style="color: red;">{{ error }}</span>
+          {% endfor %}
+      </div>
+
+      <div>
+          {{ form.perspective_Form.label }}
+          {{ form.perspective_Form(class="newCampaignForm") }}
+          {% for error in form.perspective_Form.errors %}
+              <span style="color: red;">{{ error }}</span>
+          {% endfor %}
+      </div>
+
+      <div>
+          {{ form.submit(class="btn btn-primary") }}
+      </div>
+  </form>
+
+</body>
+</html>
