+++ app/templates/viewCampaigns.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>View Campaigns</title>
+</head>
+<body>
+    <h1>View Campaigns</h1>
+    <table border="1">
+        <thead>
+            <tr>
+                <th>Name</th>
+                <th>Creation Date</th>
+                <th>Links</th>
+                <th>Summarization</th>
+                <th>Perspective</th>
+                <th>Text Generated</th>
+                <th>Image Prompt</th>
+                <th>Image Generated</th>
+            </tr>
+        </thead>
+        <tbody>
+            {% for campaign in campaigns %}
+            <tr>
+                <td>{{ campaign[0] }}</td>
+                <td>{{ campaign[1] }}</td>
+                <td>{{ campaign[2] }}</td>
+                <td>{{ campaign[3] }}</td>
+                <td>{{ campaign[4] }}</td>
+                <td>{{ campaign[5] }}</td>
+                <td>{{ campaign[6] }}</td>
+                <td>{{ campaign[7] }}</td>
+            </tr>
+            {% endfor %}
+        </tbody>
+    </table>
+</body>
+</html>
